Vehicle Management System
This program allows users to manage their vehicle's information by
creating a vehicle and its related information. Displaying a user's list
of vehicles, and copying over a list of vehicles are all functionalities
in this system. It also applies dynamic casting and polymorphism, for
example, through the various vehicles such as cars, bicycles, etc.